requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
beautifultable==0.7.0
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
Deprecated==1.2.5
idna==2.8
importlib-metadata==0.17
mock==3.0.5
more-itertools==7.0.0
packaging==19.0
pluggy==0.12.0
py==1.8.0
pycodestyle==2.5.0
PyGithub==1.43.7
PyJWT==1.7.1
pyparsing==2.4.0
pytest==4.6.2
pytest-cov==2.7.1
PyYAML==5.1
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.4
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.1
zipp==0.5.1
